requirements.txt
aiohttp
aiofiles
aioshutil
anytree
aria2p
asyncio
beautifulsoup4
cloudscraper
dnspython
flask
gevent
gunicorn
git+https://github.com/zevtyardt/lk21.git
httpx
lxml
motor
mutagen
natsort
pillow
psutil
pybase64
pymongo
pyrogram==2.0.77
python-dotenv
python-magic
requests
telegraph
tenacity
tgcrypto
uvloop
xattr
yt-dlp